requirements.txt
django-safedelete==0.5.4
djangorestframework==3.11.0
gunicorn==20.0.4
django-heroku==0.3.1